Change hash type from size_t to uint64_t for mac compatibility

+++ Bloom/HashAgnosticCascadingBloom.h
@@ -91,7 +91,7 @@ class HashAgnosticCascadingBloom
 	 * Return true if the element with the given hash values
 	 * has count >= levels.
 	 */
-	bool contains(const std::vector<size_t>& hashes) const
+	bool contains(const std::vector<uint64_t>& hashes) const
 	{
 		assert(m_data.back() != NULL);
 		return m_data.back()->contains(hashes);
@@ -101,14 +101,14 @@ class HashAgnosticCascadingBloom
 	 * Return true if the element with the given hash values
 	 * has count >= levels.
 	 */
-	bool contains(const size_t hashes[]) const
+	bool contains(const uint64_t hashes[]) const
 	{
 		assert(m_data.back() != NULL);
 		return m_data.back()->contains(hashes);
 	}
 
 	/** Add the object with the specified index to this multiset. */
-	void insert(const std::vector<size_t>& hashes)
+	void insert(const std::vector<uint64_t>& hashes)
 	{
 		for (unsigned i = 0; i < m_data.size(); ++i) {
 			assert(m_data.at(i) != NULL);
@@ -120,7 +120,7 @@ class HashAgnosticCascadingBloom
 	}
 
 	/** Add the object with the specified index to this multiset. */
-	void insert(const size_t hashes[])
+	void insert(const uint64_t hashes[])
 	{
 		for (unsigned i = 0; i < m_data.size(); ++i) {
 			assert(m_data.at(i) != NULL);
